scripts/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "bookshelf-scripts"
version = "1.0.0"
description = "Scripts to generate Bookshelf's files and for Github pipeline."
authors = [
  { name = "Gunivers", email = "contact@gunivers.net" }
]
license = {file = "LICENSE" }
dependencies = [
  "nbtlib==1.12.1",
  "pyyaml",
  "requests",
  "Jinja2"
]

[tool.setuptools.packages.find]
where = ["src"]

[options]
python_requires = ">=3.12"
